+++ src/Serializer/LeagueSerializer.php
@@ -15,4 +15,9 @@ class LeagueSerializer extends AbstractSerializer
     {
         return ['countrys', 'leagues'];
     }
+
+    protected function endpointReturnsNull(): bool
+    {
+        return true;
+    }
 }