+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Don't reuse clients and endpoints in DependencyContainer
+- `idAPIfootball` in class `NklKst\TheSportsDb\Entity\Team` is now nullable
 
 ### Removed
 - API version configuration support