+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - To be released
+
+### Fixed
+- League serializer should handle `null` response if ID was not found
+- Event serializer should handle `null` response if ID was not found
+- Skip schedule tests at the beginning and ending of the season
+
 ## [1.0.1] - 2021-05-04
 
 ### Fixed